requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.2
google-api-python-client==2.118.0
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
python-jose[cryptography]==3.3.0
supabase==2.10.0
slowapi==0.1.9
structlog>=24.1.0
cachetools>=5.3.0
